Added validation(Pet), VO FullName, config file correction

+++ src/Kind_heart.Domain/ValueObjects/Experience.cs
@@ -4,14 +4,14 @@ namespace Kind_heart.Domain.ValueObjects;
 
 public record Experience
 {
-    public float Value { get; }
+    public int Value { get; }
 
-    private  Experience(float value)
+    private  Experience(int value)
     {
         Value = value;
     }
 
-    public static Result<Experience> Create(float value)
+    public static Result<Experience> Create(int value)
     {
         if(value < 0)
             return Result.Failure<Experience>("Experience can not be less then 0");
